requirements.txt
albumentations==1.4.18
comet_ml==3.47.0
coremltools==8.0
Flask==3.0.3
GitPython==3.1.40
GitPython==3.1.43
ipython==8.12.3
matplotlib==3.8.2
mss==9.0.2
nncf==2.13.0
numpy==2.1.2
onnx==1.16.0
onnxruntime==1.19.2
onnxsim==0.4.36
opencv_python==4.8.1.78
openvino==2024.4.0
paddle==1.0.2
pafy==0.5.5
pandas==2.2.3
Pillow==10.4.0
psutil==5.9.6
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
setuptools==68.0.0
tensorflow==2.17.0
tensorflowjs==4.21.0
tensorrt==10.5.0
tflite_runtime==2.14.0
tflite_support==0.4.4
thop==0.1.1.post2209072238
torch==2.1.0+cu121
torchvision==0.16.0+cu121
tqdm==4.66.1
train==0.0.5
tritonclient==2.50.0
ultralytics==8.1.47
x2paddle==1.5.0
